% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUcopula.R
\docType{data}
\name{natperils}
\alias{natperils}
\title{natperils}
\description{
This is the 19-dimensional data set presented in Neumann et al. (2019), Tab. 2 and Tab. 3.
}
\details{
The data contains insurance losses from a non-life portfolio of natural perils in 19 areas in central Europe over a time period of 20 years. The monetary unit is 1 million €.
}
\references{
Neumann, A., Bodnar, T., Pfeifer, D., & Dickhaus, T. (2019). Multivariate multiple test procedures based on nonparametric copula estimation. Biometrical Journal, 61(1), 40-61.
}
\keyword{insurance}
\keyword{losses}
\keyword{natural}
\keyword{non-life}
\keyword{perils,}
