% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUcopula.R
\docType{data}
\name{stormflood}
\alias{stormflood}
\title{storm and flood losses data set}
\description{
This is the data set used in Cottin, Pfeifer (2014).
    The table contains some original data from an insurance portfolio of storm and flooding losses, observed over a period of 20 years.
}
\references{
Cottin, Claudia, and Dietmar Pfeifer. "From Bernstein polynomials to Bernstein copulas." J. Appl. Funct. Anal 9.3-4 (2014): 277-288.
}
\keyword{and}
\keyword{flood}
\keyword{losses}
\keyword{storm}
